Safari-609.1.11.1/LayoutTests/js/dfg-proven-sqrt-backwards-propagation-expected.txt

Tests that the DFG knows that a Math.sqrt could potentially use value in arbitrary ways, and not just in a context that converts values to numbers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is Math.sqrt(1.5)
PASS foo(array, 0) is void 0
PASS foo(array, 0) is void 0
PASS foo(array, 0) is void 0
PASS foo(array, 0) is void 0
PASS foo(array, 0) is void 0
PASS foo(array, 0) is void 0
PASS foo(array, 0) is void 0
PASS foo(array, 0) is void 0
PASS foo(array, 0) is void 0
PASS foo(array, 0) is void 0
PASS successfullyParsed is true

TEST COMPLETE
